refactor: shared 내 코드 컨벤션 관련 수정 (#ATR-308)
