+++ apps/service/entities/user-article/model/type/newsletter.ts
@@ -1,10 +1,10 @@
-import { NewsletterCategory, UploadDay } from '@/shared/type'
+import { NewsletterCategory } from '@/shared/type'
 
 export interface Newsletter {
   id: string | number
   name: string
   description: string
-  uploadDays: UploadDay[]
+  uploadDays: string
   category: NewsletterCategory
   mainLink: string
   subscribeLink: string
