+++ apps/service/shared/constant/index.ts
@@ -1,3 +1,3 @@
 export { default as NEWSLETTER_CATEGORY } from './newsletterCategory'
-export { default as UPLOAD_DAYS } from './uploadDays'
 export { default as ADMIN_EMAIL } from './adminEmail'
+export * from './googleOAuthURL'
