--- apps/service/shared/constant/uploadDays.ts
@@ -1,11 +1,0 @@
-const UPLOAD_DAYS = Object.freeze({
-  MON: '월',
-  TUE: '화',
-  WED: '수',
-  THU: '목',
-  FRI: '금',
-  SAT: '토',
-  SUN: '일',
-})
-
-export default UPLOAD_DAYS
