+++ apps/service/shared/type/index.ts
@@ -1,8 +1,6 @@
 export type { Pagination } from './pagination'
 
 export type {
-  UploadDay,
-  UploadDayName,
   NewsletterCategory,
   NewsletterCategoryName,
 } from './newsletterCategory'
