+++ apps/service/shared/ui/ContentContainer.tsx
@@ -1,9 +1,9 @@
 import React, { ReactNode } from 'react'
 
-interface CardContentProps {
+interface ContentContainerProps {
   children: ReactNode
 }
 
-export default function CardContent({ children }: CardContentProps) {
+export default function ContentContainer({ children }: ContentContainerProps) {
   return <div className="mt-3 flex gap-2">{children}</div>
 }
