+++ apps/service/shared/ui/ContentTitle.tsx
@@ -1,7 +1,11 @@
 import React from 'react'
-import { TitleType } from '../model'
 
-export default function ContentTitle({ type, content }: TitleType) {
+interface ContentTitleProps {
+  type: 'main' | 'sub'
+  content: string
+}
+
+export default function ContentTitle({ type, content }: ContentTitleProps) {
   return (
     <div
       className={`${type === 'main' ? 'font-medium' : 'text-sm text-gray-500 dark:text-gray-400'} line-clamp-2 self-center`}>
