+++ apps/service/shared/ui/ImageBox.tsx
