--- apps/service/shared/ui/content/index.ts
@@ -1,2 +1,0 @@
-export { default as ContentTitle } from './ui/ContentTitle'
-export { default as ContentContainer } from './ui/ContentContainer'
