--- apps/service/shared/ui/content/model/index.ts
@@ -1,1 +1,0 @@
-export * from './types/title'
