--- apps/service/shared/ui/content/model/types/title.ts
@@ -1,6 +1,0 @@
-interface TitleType {
-  type: 'main' | 'sub'
-  content: string
-}
-
-export type { TitleType }
