--- apps/service/shared/ui/image-box/index.ts
@@ -1,1 +1,0 @@
-export * from './ui'
