--- apps/service/shared/ui/image-box/ui/index.tsx
@@ -1,1 +1,0 @@
-export { default as ImageBox } from './ImageBox'
