+++ apps/service/shared/ui/index.ts
@@ -1,7 +1,9 @@
-export * from './content'
-export * from './image-box'
 export * from './carousel'
-export * from './title'
+export { default as Title } from './Title'
+export { default as ImageBox } from './ImageBox'
+export { default as ContentContainer } from './ContentContainer'
+export { default as ContentTitle } from './ContentTitle'
+export { default as GoogleOAuthButton } from './GoogleOAuthButton'
 export { default as LoadingSpinner } from './LoadingSpinner'
 export { default as BackBtn } from './BackBtn'
 export { default as GuideTxt } from './GuideTxt'
