--- apps/service/shared/ui/oauth-button/ui/index.ts
@@ -1,1 +1,0 @@
-export { default as GoogleOAuthButton } from './GoogleOAuthButton'
