--- apps/service/shared/ui/title/index.ts
@@ -1,1 +1,0 @@
-export { default as Title } from './Title'
